% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeContinuousUsingMeanLimitsOfCI.R
\name{AnalyzeContinuousUsingMeanLimitsOfCI}
\alias{AnalyzeContinuousUsingMeanLimitsOfCI}
\title{Analyze using a simplified limits of confidence interval design}
\usage{
AnalyzeContinuousUsingMeanLimitsOfCI(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Data frame which consists of data generated in current simulation.}

\item{DesignParam}{List of Design and Simulation Parameters required to perform analysis.}

\item{LookInfo}{A list containing input parameters related to multiple looks, which the user may need to compute 
test statistics and perform tests. Users should access the variables using their names 
(e.g., `LookInfo$NumLooks`) rather than by their order. Important variables in group sequential designs include:

- `LookInfo$NumLooks`: An integer representing the number of looks in the study.
- `LookInfo$CurrLookIndex`: An integer representing the current index look, starting from 1.
- `LookInfo$CumEvents`: A vector of length `LookInfo$NumLooks`, containing the cumulative number of events at each look.
- `LookInfo$RejType`: A code representing rejection types. Possible values are:
  - **Efficacy Only:**
    - `0`: 1-Sided Efficacy Upper.
    - `2`: 1-Sided Efficacy Lower.
  - **Futility Only:**
    - `1`: 1-Sided Futility Upper.
    - `3`: 1-Sided Futility Lower.
  - **Efficacy and Futility:**
    - `4`: 1-Sided Efficacy Upper and Futility Lower.
    - `5`: 1-Sided Efficacy Lower and Futility Upper.}

\item{UserParam}{A list of user defined parameters in East or East Horizon. UserParam must be supplied, the list must contain the following named elements:
\describe{
  \item{UserParam$dMAV}{A value (-Inf, Inf) that specifics the Minimum Acceptable Value (MAV).}
  \item{UserParam$dTV}{A value (-Inf, Inf) that specifies the Target Value (TV).}
  \item{UserParam$dConfLevel}{A value (0,1) that specifies the confidence level for the t.test() function in base R library.}
}}
}
\value{
TestStat A double value of the computed test statistic

Decision An integer value: Decision = 0 --> No boundary crossed
                                   Decision = 1 --> Lower Efficacy Boundary Crossed
                                   Decision = 2 --> Upper Efficacy Boundary Crossed
                                   Decision = 3 --> Futility Boundary Crossed
                                   Decision = 4 --> Equivalence Boundary Crossed

Delta The difference in the estimates, is utilized in East Horizon Explore to create the observed graph

ErrorCode An integer value:  ErrorCode = 0 --> No Error
}
\description{
In this simplified example of upper and lower confidence boundary designs, if it is likely that the treatment difference is above 
              the Minimum Acceptable Value (MAV) then a Go decision is made.  
              If a Go decision is not made, then if is is unlikely that the treatment difference is above the Target Value (TV) a No-Go decision is made.      
              In this example, the t.test() from base package in R is utilized to analyze the data and compute at user-specified confidence interval using UserParam$dConfLevel.  
              The team would like to make a Go decision if there is at least a 90% chance that the difference in treatment is greater than the MAV.  
              If a Go decision is not made, then a No-Go decision is made if there is less than a 10% chance the difference is greater than the TV.  
              Using a frequentist CI an approximation to this design can be done by the logic described below.
              At an analysis, if the Lower Limit of the CI, denoted by LL, is greater than user-specified UserParam$dMAV then a Go decision is made.  
              
              If a Go decision is not made, then if the Upper Limit of the CI, denoted by UL, is less than user-specified UserParam$TV a No-Go decision is made.  
              Specifically, 
                 if LL > UserParam$dMAV --> Go
                 if UL < UserParam$dTV --> No-Go
              Otherwise, continue to the next analysis. 
              At the Final Analysis: If LL > UserParam$dMAV  then a Go decision is made, otherwise, a No-Go decision is made
}
\note{
This function is only applicable to the case where MAV <= TV.  
      In this example, the boundary information that is computed and sent from East or East Horizon is ignored in order to implement this decision approach.
}
