% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeContinuousUsingTTestNormal.R
\name{AnalyzeContinuousUsingTTestNormal}
\alias{AnalyzeContinuousUsingTTestNormal}
\title{Analyze Continuous Data Using t-Test}
\usage{
AnalyzeContinuousUsingTTestNormal(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Data frame that contains simulated data for the current simulation.}

\item{DesignParam}{List of design and simulation parameters required to perform the analysis.}

\item{LookInfo}{A list of input parameters related to multiple looks in group sequential designs.
Variables should be accessed by names (e.g., \code{LookInfo$NumLooks}). Important variables include:
\itemize{
\item \code{LookInfo$NumLooks}: Integer, number of looks in the study.
\item \code{LookInfo$CurrLookIndex}: Integer, current look index (starting from 1).
\item \code{LookInfo$CumEvents}: Vector, cumulative number of events at each look.
\item \code{LookInfo$RejType}: Code representing rejection types. Possible values include:
\item \strong{Efficacy Only:}
\itemize{
\item \code{0}: 1-Sided Efficacy Upper.
\item \code{2}: 1-Sided Efficacy Lower.
}
\item \strong{Futility Only:}
\itemize{
\item \code{1}: 1-Sided Futility Upper.
\item \code{3}: 1-Sided Futility Lower.
}
\item \strong{Efficacy and Futility:}
\itemize{
\item \code{4}: 1-Sided Efficacy Upper and Futility Lower.
\item \code{5}: 1-Sided Efficacy Lower and Futility Upper.
}
}}

\item{UserParam}{A list of user-defined parameters in East or East Horizon. The default is \code{NULL}.
For this example, user-defined parameters are not included.}
}
\value{
A list containing the following elements:
\describe{
\item{TestStat}{A double representing the computed test statistic.}
\item{Decision}{Required integer value indicating the decision made:
\describe{
\item{0}{No boundary crossed (neither efficacy nor futility).}
\item{1}{Lower efficacy boundary crossed.}
\item{2}{Upper efficacy boundary crossed.}
\item{3}{Futility boundary crossed.}
\item{4}{Equivalence boundary crossed.}
}}
\item{ErrorCode}{Optional integer value:
\describe{
\item{0}{No error.}
\item{> 0}{Non-fatal error; current simulation is aborted but subsequent simulations continue.}
\item{< 0}{Fatal error; no further simulations are attempted.}
}}
}
}
\description{
Performs hypothesis testing using the \code{t.test()} function in base R to analyze continuous data under the assumption of a normal distribution. This function demonstrates how analysis and decision-making can be modified in a simple approach. The test statistic is compared to the upper boundary computed and sent by East as an input. Note that this example does not include a futility rule.
}
