% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeTTEUsingHazardRatioLimitsOfCI.R
\name{AnalyzeTTEUsingHazardRatioLimitsOfCI}
\alias{AnalyzeTTEUsingHazardRatioLimitsOfCI}
\title{Analyze Time-to-Event Data Using Hazard Ratio Limits of Confidence Interval}
\usage{
AnalyzeTTEUsingHazardRatioLimitsOfCI(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Data frame consisting of data generated in the current simulation.}

\item{DesignParam}{List of design and simulation parameters required to perform the analysis.}

\item{LookInfo}{A list containing input parameters related to multiple looks. Users should access variables using their names, such as:
\itemize{
\item \code{LookInfo$NumLooks}: Integer representing the number of looks in the study.
\item \code{LookInfo$CurrLookIndex}: Integer representing the current index look, starting from 1.
\item \code{LookInfo$CumEvents}: Vector of cumulative number of events at each look.
\item \code{LookInfo$RejType}: Code representing rejection types, with possible values:
\itemize{
\item \strong{Efficacy Only}:
\itemize{
\item \code{0}: 1-Sided Efficacy Upper.
\item \code{2}: 1-Sided Efficacy Lower.
}
\item \strong{Futility Only}:
\itemize{
\item \code{1}: 1-Sided Futility Upper.
\item \code{3}: 1-Sided Futility Lower.
}
\item \strong{Efficacy and Futility}:
\itemize{
\item \code{4}: 1-Sided Efficacy Upper and Futility Lower.
\item \code{5}: 1-Sided Efficacy Lower and Futility Upper.
}
}
}}

\item{UserParam}{A list of user-defined parameters. Must contain the following named elements:
\describe{
\item{UserParam$dMAV}{Numeric; specifies the Minimum Acceptable Value (MAV).}
\item{UserParam$dTV}{Numeric; specifies the Target Value (TV).}
\item{UserParam$dConfLevel}{Numeric (0,1); specifies the confidence level for the \code{t.test()} function.}
}}
}
\value{
A list containing the following elements:
\describe{
\item{HazardRatio}{A double representing the computed hazard ratio.}
\item{Decision}{Required integer value indicating the decision made:
\describe{
\item{0}{No boundary crossed (neither efficacy nor futility).}
\item{1}{Lower efficacy boundary crossed.}
\item{2}{Upper efficacy boundary crossed.}
\item{3}{Futility boundary crossed.}
\item{4}{Equivalence boundary crossed.}
}}
\item{ErrorCode}{Optional integer value:
\describe{
\item{0}{No error.}
\item{> 0}{Non-fatal error; current simulation is aborted but subsequent simulations continue.}
\item{< 0}{Fatal error; no further simulations are attempted.}
}}
}
}
\description{
This function analyzes time-to-event data using a simplified design based on upper and lower confidence boundaries for hazard ratios (HR).
It determines whether to make a "Go" or "No Go" decision by assessing the likelihood of the HR being below specified limits.
Specifically, the function utilizes the \code{coxph()} model from the survival package to estimate the log hazard ratio and its standard error.
The decision-making process is as follows:
\itemize{
\item If the upper limit (UL) of the confidence interval is below the Minimum Acceptable Value (MAV), a "Go" decision is made.
\item If the lower limit (LL) of the confidence interval is above the Target Value (TV), a "No Go" decision is made.
\item Otherwise, the analysis continues to the next look.
}

At the final analysis:
\itemize{
\item If UL < MAV, a "Go" decision is made.
\item Otherwise, a "No Go" decision is made.
}

HR and log HR are monotonically related. Since \code{coxph()} outputs results for log HR, the function uses the log HR scale for decision-making.
}
