% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineAllRFiles.R
\name{CombineAllRFiles}
\alias{CombineAllRFiles}
\title{Combine All R Files}
\usage{
CombineAllRFiles(
  strOutFileName = NA,
  strDirectory = "",
  strFileNameToIgnore = NA
)
}
\arguments{
\item{strOutFileName}{The name of the output file. If not provided, the function will return the combined content.}

\item{strDirectory}{The directory where the R files are located. Defaults to the current working directory.}

\item{strFileNameToIgnore}{The name, or part of the name, of any file to ignore. Defaults to \code{NA}.}
}
\value{
A list containing the following elements:
\item{nQtyCombinedFiles}{The number of files combined.}
\item{strCombinedContents}{The combined content of all the R files (only if strOutFileName is NA).}
\item{strReturn}{A string summarizing the operation, including the names of the combined files.}
}
\description{
This function combines the contents of all R files in a specified directory into one file.
Files are combined in alphabetical order and the output file itself is automatically excluded when it is located in the input directory.
}
\examples{
\dontrun{
  result <- CombineAllRFiles(
      strOutFileName = "combined.R",
      strDirectory = "/path/to/your/directory"
  )
  print( result$strReturn )
}

}
\seealso{
\code{\link[base]{list.files}}, \code{\link[base]{file}}, \code{\link[base]{readLines}}, \code{\link[base]{writeLines}}
}
