% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeHazardWeibull.R
\name{ComputeHazardWeibull}
\alias{ComputeHazardWeibull}
\title{Compute Hazard of Weibull Distribution}
\usage{
ComputeHazardWeibull(vTime, dShape, dScale)
}
\arguments{
\item{vTime}{A vector of times to compute the hazard of the Weibull distribution.}

\item{dShape}{The shape parameter of the Weibull distribution. See `rweibull`.}

\item{dScale}{The scale parameter of the Weibull distribution. See `rweibull`.}
}
\description{
Function to compute the hazard of the Weibull distribution.
}
