% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeScaleGivenShapeMedian.R
\name{ComputeScaleGivenShapeMedian}
\alias{ComputeScaleGivenShapeMedian}
\title{Compute Weibull scale parameter}
\usage{
ComputeScaleGivenShapeMedian(dShape, dMedian)
}
\arguments{
\item{dShape}{The shape parameter of the Weibull distribution.}

\item{dMedian}{The median of the Weibull distribution.}
}
\description{
Compute the scale parameter for the Weibull distribution
with a given median (\code{dMedian}) and shape parameter (\code{dShape}).
}
