% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCyneRgyFuntion.R
\name{CreateCyneRgyFuntion}
\alias{CreateCyneRgyFuntion}
\title{Create new CyneRgy Function using provided templates. These R function that is created can be used in connection with Cytel-R integration.}
\usage{
CreateCyneRgyFuntion(
  strFunctionType,
  strNewFunctionName = "",
  strDirectory = NA,
  bOpen = TRUE
)
}
\description{
{ Description: This function will create a new file containing the template for the desired CyneRgy function. }
}
