% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonRepeatedMeasuresFunctions.R
\name{GenRespDiffOfMeansRepMeasures}
\alias{GenRespDiffOfMeansRepMeasures}
\title{Generate Repeated Measures Responses}
\usage{
GenRespDiffOfMeansRepMeasures(
  NumSub,
  NumVisit,
  ArrivalTime,
  TreatmentID,
  Inputmethod,
  VisitTime,
  MeanControl,
  MeanTrt,
  StdDevControl,
  StdDevTrt,
  CorrMat,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{Integer number of subjects.}

\item{NumVisit}{Integer number of visits.}

\item{ArrivalTime}{Numeric subject arrival times.}

\item{TreatmentID}{Integer treatment identifiers beginning at \code{0}.}

\item{Inputmethod}{Integer identifying how response parameters are supplied.}

\item{VisitTime}{Numeric vector of visit times.}

\item{MeanControl}{Numeric control-arm means by visit.}

\item{MeanTrt}{Numeric treatment-arm means by visit.}

\item{StdDevControl}{Numeric control-arm standard deviations by visit.}

\item{StdDevTrt}{Numeric treatment-arm standard deviations by visit.}

\item{CorrMat}{Numeric within-subject correlation matrix.}

\item{UserParam}{Optional list of user-defined parameters.}
}
\value{
A list in the format required by the response integration point.
}
\description{
Calls the response-generation implementation from the common
\verb{2ArmNormalRepeatedMeasuresResponseGeneration} example. This function requires the suggested \code{MASS} package.
}
