% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitLoadCyneRgy.R
\name{InitLoadCyneRgy}
\alias{InitLoadCyneRgy}
\title{Initialize is used to load required libraries and create any global}
\usage{
InitLoadCyneRgy(Seed, UserParam = NULL)
}
\description{
Initialize is used to load required libraries and create any global
}
