% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbExpGreaterCtrlBeta.R
\name{ProbExpGreaterCtrlBeta}
\alias{ProbExpGreaterCtrlBeta}
\title{Compute Posterior Probability of Experimental Treatment Being Greater than Control}
\usage{
ProbExpGreaterCtrlBeta(
  vOutcomesS,
  vOutcomesE,
  dAlphaS,
  dBetaS,
  dAlphaE,
  dBetaE
)
}
\arguments{
\item{vOutcomesS}{A vector of binary outcomes (0 or 1) for the control treatment.}

\item{vOutcomesE}{A vector of binary outcomes (0 or 1) for the experimental treatment.}

\item{dAlphaS}{The alpha parameter of the Beta prior for the control treatment.}

\item{dBetaS}{The beta parameter of the Beta prior for the control treatment.}

\item{dAlphaE}{The alpha parameter of the Beta prior for the experimental treatment.}

\item{dBetaE}{The beta parameter of the Beta prior for the experimental treatment.}
}
\value{
A list containing:
\item{dPostProb}{The posterior probability that the success rate of the experimental treatment
is greater than that of the control treatment.}
}
\description{
Function to perform statistical analysis using a Beta-Binomial Bayesian model.
It computes the posterior probability that the success rate of an experimental treatment
exceeds that of a control treatment, based on observed outcomes.
}
\details{
In the Beta-Binomial model, it is assumed that the probability of success (\eqn{\pi}) follows a Beta distribution:
\eqn{\pi \sim Beta(\alpha, \beta)}. Given observed binary outcomes, the posterior distribution of \eqn{\pi} is:
\eqn{\pi | \text{data} \sim \text{Beta}(\alpha + \text{\# successes}, \beta + \text{\# non-successes})}.
This function samples from the posterior distributions of the success probabilities for both control and experimental treatments,
and calculates the posterior probability that the experimental treatment has a higher success rate than the control treatment.
}
