% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonRandomizationFunctions.R
\name{RandomizationSubjectsUsingSampleFunctionInR}
\alias{RandomizationSubjectsUsingSampleFunctionInR}
\title{Randomize Subjects to Two Arms Using R's \code{sample()} Function}
\usage{
RandomizationSubjectsUsingSampleFunctionInR(
  NumSub,
  NumArms,
  AllocRatio,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{Integer number of subjects to randomize.}

\item{NumArms}{Integer number of trial arms. This function supports two arms.}

\item{AllocRatio}{Numeric experimental-to-control allocation ratio.}

\item{UserParam}{Optional list of user-defined parameters.}
}
\value{
A list in the format required by the randomization integration point.
}
\description{
Calls the \code{sample()} implementation from the common \code{RandomizeSubjects} example.
}
