% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectExpUsingBayesianRule.R
\name{SelectExpUsingBayesianRule}
\alias{SelectExpUsingBayesianRule}
\title{Select Experimental Treatments Using a Bayesian Rule}
\usage{
SelectExpUsingBayesianRule(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Dataframe containing data generated in the current simulation.}

\item{DesignParam}{List of design and simulation parameters required to perform treatment selection.}

\item{LookInfo}{List containing design and simulation parameters, which might be required to perform treatment selection.}

\item{UserParam}{A list of user-defined parameters in East or East Horizon. The default is NULL.
The list must contain the following named elements:
\describe{
\item{UserParam$dPriorAlpha}{A value (0,1) defining the prior alpha parameter of the beta distribution.}
\item{UserParam$dPriorBeta}{A value (0,1) specifying the prior beta parameter of the beta distribution.}
\item{UserParam$dHistoricResponseRate}{A value (0,1) specifying the historic response rate.}
\item{UserParam$dMinPosteriorProbability}{A value (0,1) specifying the posterior probability needed to exceed the historic response rate for experimental treatment selection.}
}}
}
\value{
A list containing:
\item{TreatmentID}{A vector of experimental treatment IDs selected to advance, e.g., 1, 2, ..., number of experimental treatments.}
\item{AllocRatio}{A vector of allocation ratios for the selected treatments relative to control.}
\item{ErrorCode}{An integer indicating success or error status:
\describe{
\item{ErrorCode = 0}{No error.}
\item{ErrorCode > 0}{Nonfatal error, current simulation aborted but subsequent simulations will run.}
\item{ErrorCode < 0}{Fatal error, no further simulations attempted.}
}
}
}
\description{
This function implements the MAMS design for binary outcomes and performs treatment selection at the interim analysis (IA) using a Bayesian decision rule.
At IA, an experimental treatment is selected for stage 2 if its posterior probability of exceeding a user-specified historical response rate
(\code{UserParam$dHistoricResponseRate}) is greater than a user-defined threshold (\code{UserParam$dMinPosteriorProbability}):
\code{Pr(pj > UserParam$dHistoricResponseRate | data) > UserParam$dMinPosteriorProbability}.
If no treatment satisfies this criterion, the treatment with the highest posterior probability is selected.
All experimental arms assume the same prior distribution: \code{pj ~ Beta(UserParam$dPriorAlpha, UserParam$dPriorBeta)}.
For stage 2, selected treatments are randomized against the control arm in a 2:1 ratio (experimental:control).
}
\note{
\itemize{
\item The length of \code{TreatmentID} and \code{AllocRatio} must be the same.
\item The allocation ratio for control is always 1, and \code{AllocRatio} values are relative to this. For example, an allocation value of 2 means twice as many participants are randomized to the experimental treatment compared to control.
\item The order of \code{AllocRatio} should match \code{TreatmentID}, with corresponding elements assigned their respective allocation ratios.
\item The returned vector includes only \code{TreatmentID} values for experimental treatments. For example, \code{TreatmentID = c(0, 1, 2)} is invalid because control (\code{0}) should not be included.
\item At least one treatment and one allocation ratio must be returned.
}
}
