% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateContinuousPatientOutcomePercentAtZero.R
\name{SimulateContinuousPatientOutcomePercentAtZero}
\alias{SimulateContinuousPatientOutcomePercentAtZero}
\title{Simulate Continuous Patient Outcomes with Proportion at Zero}
\usage{
SimulateContinuousPatientOutcomePercentAtZero(
  NumSub,
  TreatmentID,
  Mean,
  StdDev,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{The number of subjects to simulate. Must be an integer value.}

\item{TreatmentID}{A vector of treatment IDs, where:
\itemize{
\item \code{0} represents Treatment 1.
\item \code{1} represents Treatment 2.
The length of \code{TreatmentID} must equal \code{NumSub}.
}}

\item{Mean}{A numeric vector of length 2 specifying the mean values for the two treatments.}

\item{StdDev}{A numeric vector of length 2 specifying the standard deviations for each treatment.}

\item{UserParam}{A list of user-defined parameters. Must contain the following named elements:
\describe{
\item{UserParam$dProbOfZeroOutcomeCtrl}{Numeric (0, 1); defines the probability that a patient has an outcome of 0 for the control (Treatment 1).}
\item{UserParam$dProbOfZeroOutcomeExp}{Numeric (0, 1); defines the probability that a patient has an outcome of 0 for the experimental (Treatment 2).}
}}
}
\value{
A list containing the following elements:
\describe{
\item{Response}{A numeric vector representing the simulated outcomes for each patient.}
\item{ErrorCode}{Optional integer value:
\describe{
\item{0}{No error.}
\item{> 0}{Non-fatal error; current simulation is aborted but subsequent simulations continue.}
\item{< 0}{Fatal error; no further simulations are attempted.}
}}
}
}
\description{
Simulates patient outcomes from a normal distribution, with a specified percentage of patients having an outcome of 0. In this example, the continuous outcome represents a patient's change from baseline. This function generates patient outcomes such that, on average:
\itemize{
\item A specified proportion of patients will have a value of 0 for the outcome, as defined by \code{UserParam}.
\item The remaining patients will have their values simulated from a normal distribution using the provided mean and standard deviation parameters.
}
}
