% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SimulateContinuousPatientOutcomePercentAtZeroBetaDist.R
\name{SimulateContinuousPatientOutcomePercentAtZeroBetaDist}
\alias{SimulateContinuousPatientOutcomePercentAtZeroBetaDist}
\title{Simulate Continuous Patient Outcomes with Probability of Zero from a Beta Distribution}
\usage{
SimulateContinuousPatientOutcomePercentAtZeroBetaDist(
  NumSub,
  TreatmentID,
  Mean,
  StdDev,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{The number of subjects to simulate. Must be an integer value.}

\item{TreatmentID}{A vector of treatment IDs, where:
\itemize{
\item \code{0} represents Treatment 1.
\item \code{1} represents Treatment 2.
The length of \code{TreatmentID} must equal \code{NumSub}.
}}

\item{Mean}{A numeric vector of length 2 specifying the mean values for the two treatments.}

\item{StdDev}{A numeric vector of length 2 specifying the standard deviations for each treatment.}

\item{UserParam}{A list of user-defined parameters. Must contain the following named elements:
\describe{
\item{UserParam$dCtrlBetaParam1}{Numeric; first parameter in the Beta distribution for the control (Treatment 1).}
\item{UserParam$dCtrlBetaParam2}{Numeric; second parameter in the Beta distribution for the control (Treatment 1).}
\item{UserParam$dExpBetaParam1}{Numeric; first parameter in the Beta distribution for the experimental (Treatment 2).}
\item{UserParam$dExpBetaParam2}{Numeric; second parameter in the Beta distribution for the experimental (Treatment 2).}
}}
}
\value{
A list containing the following elements:
\describe{
\item{Response}{A numeric vector representing the simulated outcomes for each patient.}
\item{ErrorCode}{Optional integer value:
\describe{
\item{0}{No error.}
\item{> 0}{Non-fatal error; current simulation is aborted but subsequent simulations continue.}
\item{< 0}{Fatal error; no further simulations are attempted.}
}}
}
}
\description{
Simulates patient outcomes from a normal distribution, with the probability of a zero outcome being random and sampled from a Beta distribution.
The probability of a zero outcome is determined as follows:
\itemize{
\item For the control treatment, it is sampled from a \eqn{Beta(UserParam$dCtrlBetaParam1, UserParam$dCtrlBetaParam2)} distribution.
\item For the experimental treatment, it is sampled from a \eqn{Beta(UserParam$dExpBetaParam1, UserParam$dExpBetaParam2)} distribution.
This approach incorporates variability in the unknown probability of no response.
}
}
