% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTTEPatientWeibull.R
\name{SimulateTTEPatientWeibull}
\alias{SimulateTTEPatientWeibull}
\title{Simulate TTE Patient Outcomes from a Weibull Distribution}
\usage{
SimulateTTEPatientWeibull(
  NumSub,
  NumArm,
  TreatmentID,
  SurvMethod,
  NumPrd,
  PrdTime,
  SurvParam,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{The number of patient times to generate for the trial. This is a single numeric value, e.g., 250.}

\item{NumArm}{The number of arms in the trial, a single numeric value. For a two-arm trial, this will be 2.}

\item{TreatmentID}{A vector of treatment IDs. \verb{0 = treatment 1}, \verb{1 = treatment 2}. The length of \code{TreatmentID} must be equal to \code{NumSub}.}

\item{SurvMethod}{This value is pulled from the Input Method drop-down list.
\describe{
\item{1}{Hazard Rate.}
\item{2}{Cumulative percentage survival.}
\item{3}{Medians.}
}}

\item{NumPrd}{Number of time periods that are provided.}

\item{PrdTime}{\describe{
\item{If \code{SurvMethod = 1}}{\code{PrdTime} is a vector of starting times of hazard pieces.}
\item{If \code{SurvMethod = 2}}{Times at which the cumulative percentage survivals are specified.}
\item{If \code{SurvMethod = 3}}{\code{PrdTime} is 0 by default.}
}}

\item{SurvParam}{A 2-D array of parameters to generate the survival times, depending on the table in the Response Generation tab.
\describe{
\item{If \code{SurvMethod = 1}}{\code{SurvParam} is an array (\code{NumPrd} rows, \code{NumArm} columns) that specifies arm-by-arm hazard rates
(one rate per arm per piece). Thus, \code{SurvParam[i, j]} specifies the hazard rate in the \code{i}th period for the \code{j}th arm.
Arms are in columns where column 1 is control and column 2 is experimental. Time periods are in rows, where row 1 is time period 1, row 2 is time period 2, etc.}
\item{If \code{SurvMethod = 2}}{\code{SurvParam} is an array (\code{NumPrd} rows, \code{NumArm} columns) that specifies arm-by-arm the cumulative percentage survivals
(one value per arm per piece). Thus, \code{SurvParam[i, j]} specifies the cumulative percentage survivals in the \code{i}th period for the \code{j}th arm.}
\item{If \code{SurvMethod = 3}}{\code{SurvParam} will be a \verb{1 x 2} array with median survival times for each arm. Column 1 is control, column 2 is experimental.}
}}

\item{UserParam}{A list of user-defined parameters. Must contain the following named elements:
\describe{
\item{\code{UserParam$dShapeCtrl}}{The shape parameter in the Weibull distribution for the control treatment.}
\item{\code{UserParam$dScaleCtrl}}{The scale parameter in the Weibull distribution for the control treatment.}
\item{\code{UserParam$dShapeExp}}{The shape parameter in the Weibull distribution for the experimental treatment.}
\item{\code{UserParam$dScaleExp}}{The scale parameter in the Weibull distribution for the experimental treatment.}
}}
}
\value{
A list with the following components:
\item{\code{SurvivalTime}}{A vector of simulated survival times for patients.}
\item{\code{ErrorCode}}{Optional integer value:
\describe{
\item{0}{No error.}
\item{> 0}{Non-fatal error; current simulation is aborted but subsequent simulations continue.}
\item{< 0}{Fatal error; no further simulations are attempted.}
}}
}
\description{
This function simulates patient data from a Weibull (shape, scale) distribution. The \code{rweibull} function in the \code{stats} package
is used to simulate the survival time. See help on \code{rweibull}.

The required function signature for integration with East or East Horizon includes the \code{SurvMethod}, \code{NumPrd}, \code{PrdTime}, and \code{SurvParam},
which are ignored in this function, and only the parameters in \code{UserParam} are utilized.
}
