% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeContinuousUsingMeanLimitsOfCI.R
\name{AnalyzeContinuousUsingMeanLimitsOfCI}
\alias{AnalyzeContinuousUsingMeanLimitsOfCI}
\title{Analyze Continuous Data Using Mean Limits of Confidence Interval}
\usage{
AnalyzeContinuousUsingMeanLimitsOfCI(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Data frame consisting of data generated in the current simulation.}

\item{DesignParam}{List of design and simulation parameters required to perform the analysis.}

\item{LookInfo}{A list of input parameters related to multiple looks in group sequential designs.
Variables should be accessed by names (e.g., \code{LookInfo$NumLooks}). Important variables include:
\itemize{
\item \code{LookInfo$NumLooks}: Integer, number of looks in the study.
\item \code{LookInfo$CurrLookIndex}: Integer, current look index (starting from 1).
\item \code{LookInfo$CumEvents}: Vector, cumulative number of events at each look.
\item \code{LookInfo$RejType}: Code representing rejection types. Possible values include:
\item \strong{Efficacy Only:}
\itemize{
\item \code{0}: 1-Sided Efficacy Upper.
\item \code{2}: 1-Sided Efficacy Lower.
}
\item \strong{Futility Only:}
\itemize{
\item \code{1}: 1-Sided Futility Upper.
\item \code{3}: 1-Sided Futility Lower.
}
\item \strong{Efficacy and Futility:}
\itemize{
\item \code{4}: 1-Sided Efficacy Upper and Futility Lower.
\item \code{5}: 1-Sided Efficacy Lower and Futility Upper.
}
}}

\item{UserParam}{A list of user-defined parameters. Must contain the following named elements:
\describe{
\item{UserParam$dMAV}{Numeric; specifies the Minimum Acceptable Value (MAV).}
\item{UserParam$dTV}{Numeric; specifies the Target Value (TV).}
\item{UserParam$dConfLevel}{Numeric (0,1); specifies the confidence level for the \code{t.test()} function.}
}}
}
\value{
A list containing the following elements:
\describe{
\item{TestStat}{A double representing the computed test statistic.}
\item{Decision}{Required integer value indicating the decision made:
\describe{
\item{0}{No boundary crossed (neither efficacy nor futility).}
\item{1}{Lower efficacy boundary crossed.}
\item{2}{Upper efficacy boundary crossed.}
\item{3}{Futility boundary crossed.}
\item{4}{Equivalence boundary crossed.}
}}
\item{ErrorCode}{Optional integer value:
\describe{
\item{0}{No error.}
\item{> 0}{Non-fatal error; current simulation is aborted but subsequent simulations continue.}
\item{< 0}{Fatal error; no further simulations are attempted.}
}}
\item{Delta}{Estimated difference between experimental and control treatments.}
}
}
\description{
This function performs analysis using a simplified limits of confidence interval design for continuous outcomes.
The analysis determines whether a "Go" or "No-Go" decision is made based on the lower and upper limits of
a user-specified confidence interval. It uses the \code{t.test} function from the base R library to compute the
confidence interval. The decision-making process is based on the following logic:
\itemize{
\item If the lower limit of the confidence interval (LL) is greater than the Minimum Acceptable Value (MAV), a "Go" decision is made.
\item If a "Go" decision is not made, and the upper limit of the confidence interval (UL) is less than the Target Value (TV), a "No-Go" decision is made.
\item Otherwise, continue to the next analysis.
\item At the final analysis, if LL > MAV, a "Go" decision is made; otherwise, a "No-Go" decision is made.
}

This function assumes MAV ≤ TV and ignores boundary information sent from East or East Horizon to implement this decision approach.
}
\note{
This function is applicable only when MAV ≤ TV.
}
