% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeUsingBetaBinomial.R
\name{AnalyzeUsingBetaBinomial}
\alias{AnalyzeUsingBetaBinomial}
\title{Analyze Using Bayesian Beta-Binomial Model for Binary Data
Analyze for efficacy using a beta prior to compute the posterior probability that experimental is better than standard of care.}
\usage{
AnalyzeUsingBetaBinomial(SimData, DesignParam, LookInfo, UserParam = NULL)
}
\arguments{
\item{SimData}{Data frame which consists of data generated in current simulation.}

\item{DesignParam}{List of Design and Simulation Parameters required to perform analysis.}

\item{LookInfo}{List containing Design and Simulation Parameters, which might be required to perform analysis.}

\item{UserParam}{A list of user defined parameters in East. The default must be NULL.
If UserParam is supplied, the list must contain the following named elements:
UserParam$dAlphaS
UserParam$dBetaS
UserParam$dBetaE
UserParam$dAlphaE
UserParam$dUpperCutoffEfficacy - A value (0,1) that specifies the upper cutoff for the efficacy check. Above this value will declare efficacy 
UserParam$dLowerCutoffForFutility - A value (0,1) that specified the lower cutoff for the futility check. Below this value will declare futility.}
}
\value{
TestStat A double value of the computed test statistic

Decision An integer value: Decision = 0 --> No boundary crossed
                                   Decision = 1 --> Lower Efficacy Boundary Crossed
                                   Decision = 2 --> Upper Efficacy Boundary Crossed
                                   Decision = 3 --> Futility Boundary Crossed
                                   Decision = 4 --> Equivalence Boundary Crossed

ErrorCode An integer value:  ErrorCode = 0 --> No Error
}
\description{
In this version, the analysis for efficacy is to assume a beta prior to compute the posterior probability that experimental is better than standard of care.
             The futility is based on a Bayesian predictive probability.  
             The prior for the prediction and the analysis do NOT need to be the same.  
             This function requires more info in the glDesign than the previous AnalyzeUsingBetaBinomBayesianModel
}
\note{
In this example we assume a Bayesian model and use posterior probabilities for decision making

Helpful Hints:
      There is often info that East sends to R that are not shown in a given example.  It can be very helpful to save the input 
      objects and then load them into your R session and inspect them.  This can be done with the following R code in your function.

      saveRDS( SimData,     "SimData.Rds")
      saveRDS( DesignParam, "DesignParam.Rds" )
      saveRDS( LookInfo,    "LookInfo.Rds" )

      The above code will save each of the input objects to a file so they may be examined within R.
TODO(Kyle): I am not sure how to define the alpha and beta user parameters. Could you define and add to documentation?
TODO(Kyle): Should the functions at the bottom be left at the bottom or do they need to be added to top documentation?
}
