% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeUsingBetaBinomial.R
\name{AnalyzeUsingBetaBinomial}
\alias{AnalyzeUsingBetaBinomial}
\title{Analyze Using Bayesian Beta-Binomial Model for Binary Data
Analyze for efficacy using a beta prior to compute the posterior probability that experimental is better than standard of care.}
\usage{
AnalyzeUsingBetaBinomial(SimData, DesignParam, LookInfo, UserParam = NULL)
}
\arguments{
\item{SimData}{Data frame which consists of data generated in current simulation.}

\item{DesignParam}{List of Design and Simulation Parameters required to perform analysis.}

\item{LookInfo}{List containing Design and Simulation Parameters, which might be required to perform analysis.}

\item{UserParam}{A list of user defined parameters in East. The default must be NULL.
If UseParam is supplied, the list must contain the following named elements:
\describe{
   \item{UserParam$dAlphaS}{}
   \item{UserParam$dBetaS}{}
   \item{UserParam$dAlphaE}{}
   \item{UserParam$dBetaE}{}
   \item{UserParam$dMAV}{}
   \item{UserParam$dUpperCutoffEfficacy}{ A value (0,1) that specifies the upper cutoff for the efficacy check. Above this value will declare efficacy }
   \item{UserParam$dLowerCutoffForFutility}{A value (0,1) that specified the lower cutoff for the futility check. Below this value will declare futility. }
   }}
}
\value{
TestStat A double value of the computed test statistic

Decision An integer value: Decision = 0 --> No boundary crossed
                                   Decision = 1 --> Lower Efficacy Boundary Crossed
                                   Decision = 2 --> Upper Efficacy Boundary Crossed
                                   Decision = 3 --> Futility Boundary Crossed
                                   Decision = 4 --> Equivalence Boundary Crossed

ErrorCode An integer value:  ErrorCode = 0 --> No Error
}
\description{
In this version, the analysis for efficacy is to assume a beta prior to compute the posterior probability that experimental is better than standard of care.
             The futility is based on a Bayesian predictive probability.  
             The prior for the prediction and the analysis do NOT need to be the same.  
             This function requires more info in the glDesign than the previous AnalyzeUsingBetaBinomBayesianModel
}
\note{
In this example we assume a Bayesian model and use posterior probabilities for decision making

When using simulation to obtain the frequentist Operating Characteristic (OC) of a Bayesian design, you should set dLowerCutoffForFutility = 0
when simulating under the null case in order to obtain the false-positive rate of the non-binding futility rule.  
When you set dLowerCutoffForFutility > 0, simulation will provide the OC of the binding futility rule because the rule is ALWAYS followed.

Helpful Hints:
      There is often info that East sends to R that are not shown in a given example.  It can be very helpful to save the input 
      objects and then load them into your R session and inspect them.  This can be done with the following R code in your function.

      saveRDS( SimData,     "SimData.Rds")
      saveRDS( DesignParam, "DesignParam.Rds" )
      saveRDS( LookInfo,    "LookInfo.Rds" )

      The above code will save each of the input objects to a file so they may be examined within R.
TODO(Kyle): I am not sure how to define the alpha and beta user parameters. Could you define and add to documentation?
}
