% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeUsingHazardRatioLimitsOfCI.R
\name{AnalyzeUsingHazardRatioLimitsOfCI}
\alias{AnalyzeUsingHazardRatioLimitsOfCI}
\title{Analyze using a simplified limits of confidence interval design}
\usage{
AnalyzeUsingHazardRatioLimitsOfCI(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Data frame which consists of data generated in current simulation.}

\item{DesignParam}{List of Design and Simulation Parameters required to perform analysis.}

\item{LookInfo}{List containing Design and Simulation Parameters, which might be required to perform analysis.}

\item{UserParam}{A list of user defined parameters in East or Solara. UserParam must be supplied, the list must contain the following named elements:
\describe{
  \item{UserParam$dMAV}{A value (0, Inf) that specifics the lower limit, eg  Minimum Acceptable Value (MAV). }
  \item{UserParam$dTV}{A value (0 Inf) that specifies the upper limit for the confidence interval, eg Target Value (TV).}
  \item{UserParam$dConfLevel}{A value (0,1) that specifies the confidence level for the t.test() function in base R library.}
}}
}
\value{
Hazard Ratio :  A double value of the computed or observed Hazard ratio

Decision An integer value: Decision = 0 --> No boundary crossed
                                   Decision = 1 --> Lower Efficacy Boundary Crossed
                                   Decision = 2 --> Upper Efficacy Boundary Crossed
                                   Decision = 3 --> Futility Boundary Crossed
                                   Decision = 4 --> Equivalence Boundary Crossed

ErrorCode An integer value:  ErrorCode = 0 --> No Error
}
\description{
In this simplified example of upper and lower confidence boundary designs, if it is likely that the HR (Hazard Ratio) is below the Minimum Acceptable Value (MAV) then a Go decision is made.  
              If a Go decision is not made, then if it is unlikely that the Hazard ratio is below the Target Value (TV) a No Go decision is made.      
              In this example, the coxph() from survival package in R is utilized to analyze the data and compute estimate of log HR and Std error of log HR. 
              The team would like to make a Go decision if there is at least a 90% chance that HR is below than the MAV.  
              If a Go decision is not made, then a No Go decision is made if there is less than a 10% chance the HR is less than the TV.  
         
              Specifically, if user provides upper and lower limit in Hazard ratio scale then,
              1. For Hazard Ratio 
                 if UL < UserParam$dMAV --> Go 
                 if LL > UserParam$dTV --> No Go
                 Otherwise, continue to the next analysis. 
                 
             2. For log Hazard Ratio 
                 if UL < UserParam$dMAV --> Go 
                 if LL > UserParam$dTV --> No Go
              Otherwise, continue to the next analysis. 
 Note - HR and log HR are monotonically related.
 In coxph() function, we get the analysis for log HR and hence we make the use of 2) in decision making. 
              
 At the Final Analysis: If UL < UserParam$dMAV  then a Go decision is made, otherwise, a No Go decision is made.
}
\details{
Analyze Using Hazard Ratio Limits Of CI
}
\note{
In this example, the boundary information that is computed and sent from East is ignored in order to implement this decision approach.

Helpful Hints:
      There is often info that East sends to R that are not shown in a given example.  It can be very helpful to save the input 
      objects and then load them into your R session and inspect them.  This can be done with the following R code in your function.

      saveRDS( SimData,     "SimData.Rds")
      saveRDS( DesignParam, "DesignParam.Rds" )
      saveRDS( LookInfo,    "LookInfo.Rds" )

      The above code will save each of the input objects to a file so they may be examined within R.
}
