% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeUsingMeanLimitsOfCI.R
\name{AnalyzeUsingMeanLimitsOfCI}
\alias{AnalyzeUsingMeanLimitsOfCI}
\title{Analyze using a simplified limits of confidence interval design}
\usage{
AnalyzeUsingMeanLimitsOfCI(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{Data frame which consists of data generated in current simulation.}

\item{DesignParam}{List of Design and Simulation Parameters required to perform analysis.}

\item{LookInfo}{List containing Design and Simulation Parameters, which might be required to perform analysis.}

\item{UserParam}{A list of user defined parameters in East or Solara. UserParam must be supplied, the list must contain the following named elements:
\describe{
  \item{UserParam$dLowerLimit}{A value (-Inf, Inf) that specifics the lower limit, eg  Minimum Acceptable Value (MAV). }
  \item{UserParam$dUpperLimit}{A value (-Inf, Inf) that specifies the upper limit for the confidence interval, eg Target Value (TV).}
  \item{UserParam$dConfLevel}{A value (0,1) that specifies the confidence level for the t.test() function in base R library.}
}}
}
\value{
TestStat A double value of the computed test statistic

Decision An integer value: Decision = 0 --> No boundary crossed
                                   Decision = 1 --> Lower Efficacy Boundary Crossed
                                   Decision = 2 --> Upper Efficacy Boundary Crossed
                                   Decision = 3 --> Futility Boundary Crossed
                                   Decision = 4 --> Equivalence Boundary Crossed

Delta The difference in the estimates, is utilzied in Solara to create the observed graph

ErrorCode An integer value:  ErrorCode = 0 --> No Error
}
\description{
In this simplified example of upper and lower confidence boundary designs, if it is likely that the treatment difference is above the Minimum Acceptable Value (MAV) then a Go decision is made.  
              If a Go decision is not made, then if is is unlikely that the treatment difference is above the Target Value (TV) a No Go decision is made.      
              In this example, the t.test() from base package in R is utilized to analyze the data and compute at user-specified confidence interval (dConfLevel).  
              The team would like to make a Go decision if there is at least a 90% chance that the difference in treatment is greater than the MAV.  
              If a Go decision is not made, then a No Go decision is made if there is less than a 10% chance the difference is greater than the TV.  
              Using a frequentist CI an approximation to this design can be done by the logic described below.
              At an analysis, if the Lower Limit of the CI, denoted by LL, is greater than user-specified dLowerLimit then a Go decision is made.  
              
              If a Go decision is not made, then if the Upper Limit of the CI, denoted by UL, is less than user-specified dUpperLimit a No Go decision is made.  
              Specifically, 
                 if LL > UserParam$dLowerLimit --> Go
                 if UL < UserParam$dUpperLimit --> No Go
              Otherwise, continue to the next analysis. 
              At the Final Analysis: If LL > UserParam$dLowerLimit  then a Go decision is made, otherwise, a No Go decision is made
}
\details{
Analyze using a simplified limits of confidence interval design
}
\note{
In this example, the boundary information that is computed and sent from East is ignored in order to implement this decision approach.

Helpful Hints:
      There is often info that East sends to R that are not shown in a given example.  It can be very helpful to save the input 
      objects and then load them into your R session and inspect them.  This can be done with the following R code in your function.

      saveRDS( SimData,     "SimData.Rds")
      saveRDS( DesignParam, "DesignParam.Rds" )
      saveRDS( LookInfo,    "LookInfo.Rds" )

      The above code will save each of the input objects to a file so they may be examined within R.
}
