% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineAllRFiles.R
\name{CombineAllRFiles}
\alias{CombineAllRFiles}
\title{CombineAllRFiles}
\usage{
CombineAllRFiles(strOutFileName, strDirectory = "")
}
\description{
{ Description: This function is used to combine all .R files in a directory into a single file for use in Cytel products.  }
}
