% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctionsWeibull.R
\name{ComputeScaleGivenShapeMedian}
\alias{ComputeScaleGivenShapeMedian}
\title{Compute the scale parameter for the Weibull distribution with median = dMedian and scale parameter = dScale}
\usage{
ComputeScaleGivenShapeMedian(dShape, dMedian)
}
\arguments{
\item{dShape}{The shape of the Weibull distribution}

\item{dMedian}{The median of the Weibull distribution}
}
\description{
Function to compute the scale given the shape and median
}
