% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCyneRgyExample.R
\name{CreateCyneRgyExample}
\alias{CreateCyneRgyExample}
\title{Create new CyneRgy example using provided templates. The directory created can be used in connection with Cytel-R integration.}
\usage{
CreateCyneRgyExample(
  strFunctionType,
  strNewExampleName = "",
  strDirectory = NA
)
}
\description{
{ Description: This function will create a new directory containing the necessary files for the desired CyneRgy template. }
}
