% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCyneRgyFunction.R
\name{CreateCyneRgyFunction}
\alias{CreateCyneRgyFunction}
\title{Create a new CyneRgy function using provided templates that can be utilized with R integration points in Cytel products.}
\usage{
CreateCyneRgyFunction(
  strFunctionType = "",
  strNewFunctionName = NA,
  strDirectory = NA,
  bOpen = TRUE
)
}
\arguments{
\item{strFunctionType}{A string variable that provides the function type to create.   This argument is used to determine which integration template should be used.}

\item{strNewFunctionName}{The name of the new function to create.   This argument is also used to name the file.  The resulting file is named [strNewFunctionName].R}

\item{strDirectory}{A sub-directory with this name is created in the current working directory.  If no value is provided, the file is created in the working directory.}

\item{bOpen}{TRUE/FALSE value, when TRUE the file is opened and when FALSE the file is not opened, just created. Note, the R Studio API is used to open the file and only works in R Studio.}
}
\description{
{Create a new CyneRgy function using provided templates that can be utilized with R integration points in Cytel products.}
}
\examples{
\dontrun{
CreateCyneRgyFunction()  # A full list of options for strFunctionType is provided

# Using the Analyze.Binary function template a new file named NewBinaryAnalysis.R is created in the working directory with a function that is ready to
# be used in places where binary data is generated in simulations. 
CreateCyneRgyFunction( "Analyze.Binary", "NewBinaryAnalysis" )  
}
}
