% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDecision.R
\name{GetDecision}
\alias{GetDecision}
\title{GetDecision}
\usage{
GetDecision(strDecision, DesignParam, LookInfo)
}
\arguments{
\item{strDecision}{is a  string with either "Efficacy", "Futility" or "Continue"}

\item{DesignParam}{This is the DesignParam sent from East Horizon Explore to the R integration for analysis.}

\item{LookInfo}{The LookInfo parameter sent from East Horizon Explore to the R integration for analysis.}
}
\description{
{ Description: This function takes a string for the desired decision, design and look info and return the correct decision value.  
If LookInfo is not Null then looking at LookInfo$RejType can help determine the design type
  LookInfo$RejType Codes:
    Efficacy Only:
      1 Sided Efficacy Upper = 0
      1 Sided Efficacy Lower = 2
    
    Futility Only:
      1 Sided Futility Upper = 1
      1 Sided Futility Lower = 3
    
    Efficacy Futility:
      1 Sided Efficacy Upper Futility Lower = 4 
      1 Sided Efficacy Lower Futility Upper = 5
    
    Not in East Horizon Explore Yet:
      2 Sided Efficacy Only = 6
      2 Sided Futility Only = 7
      2 Sided Efficacy Futility = 8
      Equivalence = 9

  Then using DesignParam$TailType
      0: Left Tailed
      1: Right Tailed



}
}
