% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDecisionString.R
\name{GetDecisionString}
\alias{GetDecisionString}
\title{Generate Decision String Based on Interim and Final Analysis Conditions}
\usage{
GetDecisionString(
  LookInfo,
  nLookIndex,
  nQtyOfLooks,
  bIAEfficacyCondition = FALSE,
  bIAFutilityCondition = FALSE,
  bFAEfficacyCondition = FALSE,
  bFAFutilityCondition = FALSE
)
}
\arguments{
\item{LookInfo}{List containing look information passed from East Horizon Explore to the R integration for analysis.}

\item{nLookIndex}{Integer indicating the current look index, created by the user in the analysis code.}

\item{nQtyOfLooks}{Integer indicating the total number of looks in the study, created by the user in the analysis code.}

\item{bIAEfficacyCondition}{Logical condition evaluated to determine interim efficacy at a look (defaults to \code{FALSE}).}

\item{bIAFutilityCondition}{Logical condition evaluated to determine interim futility at a look (defaults to \code{FALSE}).}

\item{bFAEfficacyCondition}{Logical condition evaluated to determine final efficacy at the last look (defaults to \code{FALSE}).}

\item{bFAFutilityCondition}{Logical condition evaluated to determine final futility at the last look (defaults to \code{FALSE}).}
}
\description{
This function evaluates look information, efficacy conditions, and futility conditions to generate the decision string (\code{strDecision}) required for the \code{GetDecision} function.
If \code{LookInfo} is not \code{NULL}, the \code{LookInfo$RejType} parameter can be used to determine the design type.

\code{LookInfo$RejType} Codes:
\itemize{
\item \strong{Efficacy Only}:
\itemize{
\item 1-Sided Efficacy Upper = 0
\item 1-Sided Efficacy Lower = 2
}
\item \strong{Futility Only}:
\itemize{
\item 1-Sided Futility Upper = 1
\item 1-Sided Futility Lower = 3
}
\item \strong{Efficacy and Futility}:
\itemize{
\item 1-Sided Efficacy Upper and Futility Lower = 4
\item 1-Sided Efficacy Lower and Futility Upper = 5
}
\item \strong{Not in East Horizon Explore Yet}:
\itemize{
\item 2-Sided Efficacy Only = 6
\item 2-Sided Futility Only = 7
\item 2-Sided Efficacy and Futility = 8
\item Equivalence = 9
}
}
}
