% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDecisionString.R
\name{GetDecisionString}
\alias{GetDecisionString}
\title{GetDecisionString}
\usage{
GetDecisionString(
  LookInfo,
  nLookIndex,
  nQtyOfLooks,
  bIAEfficacyCondition = FALSE,
  bIAFutilityCondition = FALSE,
  bFAEfficacyCondition = FALSE,
  bFAFutilityCondition = FALSE
)
}
\arguments{
\item{LookInfo}{The LookInfo parameter sent from East Horizon Explore to the R integration for analysis.}

\item{nLookIndex}{is an integer indicating the current look. This is created by the user in the analysis code.}

\item{nQtyOfLooks}{is an integer indicating the total number of looks in the study. This is created by the user in the analysis code.}

\item{bIAEfficacyCondition}{is a condition that renders to a boolean. If available and applicable, this condition is checked to declare efficacy at an interim look.}

\item{bIAfutilityCondition}{is a condition that renders to a boolean. If available and applicable, this condition is checked to declare futility at an interim look.}

\item{bFAefficacyCondition}{is a condition that renders to a boolean. If available and applicable, this condition is checked to declare efficacy at the final look.}

\item{bFAfutilityCondition}{is a condition that renders to a boolean. If available and applicable, this condition is checked to declare futility at the final look.}
}
\description{
{ Description: This function takes look information and efficacy and futility conditions to be checked and returns the strDecision needed for the GetDecision function.  
If LookInfo is not Null then looking at LookInfo$RejType can help determine the design type
  LookInfo$RejType Codes:
    Efficacy Only:
      1 Sided Efficacy Upper = 0
      1 Sided Efficacy Lower = 2
    
    Futility Only:
      1 Sided Futility Upper = 1
      1 Sided Futility Lower = 3
    
    Efficacy Futility:
      1 Sided Efficacy Upper Futility Lower = 4 
      1 Sided Efficacy Lower Futility Upper = 5
    
    Not in East Horizon Explore Yet:
      2 Sided Efficacy Only = 6
      2 Sided Futility Only = 7
      2 Sided Efficacy Futility = 8
      Equivalence = 9


}
}
