% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitLoadCyneRgy.R
\name{InitLoadCyneRgy}
\alias{InitLoadCyneRgy}
\title{Initialize CyneRgy Library for Simulations}
\usage{
InitLoadCyneRgy(Seed, UserParam = NULL)
}
\arguments{
\item{Seed}{An integer value to set the seed used for generating random numbers in R. Default is NULL.}

\item{UserParam}{A named list to pass custom scalar variables defined by users. Variables can be accessed using names,
e.g., \code{UserParam$Var1}. Default is NULL.}
}
\value{
A list containing:
\item{ErrorCode}{An integer indicating success or error status:
\describe{
\item{ErrorCode = 0}{No error.}
\item{ErrorCode > 0}{Nonfatal error, current simulation aborted but subsequent simulations will run.}
\item{ErrorCode < 0}{Fatal error, no further simulations attempted.}
}
}
}
\description{
The Initialize function initializes the R environment for all simulations. It is optional and is executed before any user-defined functions.
Key functionalities include:
\itemize{
\item Setting a seed for random number generation
\item Loading required packages
}

Here, the CyneRgy library is initialized.
}
\note{
Do not use \code{install.packages} or attempt to install new R packages in East Horizon, as this will fail. Please contact support to install libraries.
}
