% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceTagsInFile.R
\name{ReplaceTagsInFile}
\alias{ReplaceTagsInFile}
\title{ReplaceTagsInFile}
\usage{
ReplaceTagsInFile(strFileName, vTags, vReplace)
}
\description{
{ Description: This function is used to replace {{tags}} in template files. }
}
