% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceTagsInFile.R
\name{ReplaceTagsInFile}
\alias{ReplaceTagsInFile}
\title{Replace tags in a file.}
\usage{
ReplaceTagsInFile(strFileName, vTags, vReplace)
}
\arguments{
\item{strFileName}{The name of the file to use as input.  In tags, defined by {{tags}}, will be replaced with the corresponding values.}

\item{vTags}{Vector of tag names, eg FUNCTION_NAME, VARIABLE_NAME that will be replaced with the values in vReplace}

\item{vreplace}{Vector of values to replace the tags with.}
}
\value{
A TRUE/FALSE value if the functions was successful.
}
\description{
{ Description: This function is used to replace {{tags}} in template files. }
}
\examples{
 \dontrun{
vTags    <- c("FUNCTION_NAME",  "CREATION_DATE")
vReplace <- c(strNewFunctionName, strToday)
strFileName <- "MyTemplate.R" # A file that contains {{FUNCTION_NAME}} and {{CREATION_DATE}}
ReplaceTagsInFile( strFileName, vTags, vReplace )}
}
