% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{RunExample}
\alias{RunExample}
\title{Launch Example from the CyneRgy Package}
\usage{
RunExample(strExample)
}
\description{
This function opens an example included in the CyneRgy package. A new instance of RStudio will launch and open the requested example.
To obtain a current list of available examples, call \code{CyneRgy::RunExample()}, and it will display the list.
}
\note{
The function requires RStudio to open the example projects automatically. If the RStudio API is unavailable, a manual process will be needed.
}
\examples{
\dontrun{
CyneRgy::RunExample("TreatmentSelection")
}

}
