% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{RunExample}
\alias{RunExample}
\title{RunExample}
\usage{
RunExample(example)
}
\description{
{This function is used to open the examples included in this package.  
  For a current list of example call
  CyneRgy::RunExample() and you will get a list of the available examples.

  This function opens the example in R Studio. 
  }
}
\examples{
\dontrun{{CyneRgy::RunExample( "TreatmentSelection" )
}}
}
