% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SelectSpecifiedNumberOfExpWithHighestResponses.R
\name{SelectSpecifiedNumberOfExpWithHighestResponses}
\alias{SelectSpecifiedNumberOfExpWithHighestResponses}
\title{Select Treatments with the Highest Number of Responses}
\usage{
SelectSpecifiedNumberOfExpWithHighestResponses(
  SimData,
  DesignParam,
  LookInfo = NULL,
  UserParam = NULL
)
}
\arguments{
\item{SimData}{A dataframe consisting of data generated in the current simulation.}

\item{DesignParam}{A list of design and simulation parameters required to perform treatment selection.}

\item{LookInfo}{A list containing design and simulation parameters that might be required to perform treatment selection.}

\item{UserParam}{A list of user-defined parameters in East or East Horizon. The default is \code{NULL}.
The list must contain the following named elements:
\describe{
\item{QtyOfArmsToSelect}{A value defining how many treatment arms are chosen to advance. Note this number must match the number of user-specified allocation values.}
\item{Rank1AllocationRatio}{A value specifying the allocation to the arm with the highest response.}
\item{Rank2AllocationRatio}{A value specifying the allocation to the arm with the next highest response.}
}}
}
\value{
A list containing:
\item{TreatmentID}{A vector of experimental treatment IDs selected to advance, e.g., 1, 2, ..., number of experimental treatments.}
\item{AllocRatio}{A vector of allocation ratios for the selected treatments relative to control.}
\item{ErrorCode}{An integer indicating success or error status:
\describe{
\item{ErrorCode = 0}{No error.}
\item{ErrorCode > 0}{Nonfatal error, current simulation aborted but subsequent simulations will run.}
\item{ErrorCode < 0}{Fatal error, no further simulations attempted.}
}
}
}
\description{
This function is used for the Multi-Arm Multi-Stage (MAMS) design with a binary outcome and performs treatment selection at the interim analysis (IA).
At the IA, the user-specified number of experimental treatments (\code{QtyOfArmsToSelect}) that have the largest number of responses are selected. After the IA, randomization is based on user-specified inputs: 1:\code{Rank1AllocationRatio}:\code{Rank2AllocationRatio} (control, selected experimental arm with the highest number of responses, selected experimental arm with the second highest number of responses).
}
\note{
\itemize{
\item The length of \code{TreatmentID} and \code{AllocRatio} must be the same.
\item The allocation ratio for control is always 1, and \code{AllocRatio} values are relative to this. For example, an allocation value of 2 means twice as many participants are randomized to the experimental treatment compared to control.
\item The order of \code{AllocRatio} should match \code{TreatmentID}, with corresponding elements assigned their respective allocation ratios.
\item The returned vector includes only \code{TreatmentID} values for experimental treatments. For example, \code{TreatmentID = c(0, 1, 2)} is invalid because control (\code{0}) should not be included.
\item At least one treatment and one allocation ratio must be returned.
}
}
\examples{
# Example 1: Assuming the allocation in the second part of the trial is 1:2:2 for Control:Experimental 1:Experimental 2
vSelectedTreatments <- c(1, 2)  # Experimental 1 and 2 both have an allocation ratio of 2.
vAllocationRatio    <- c(2, 2)
nErrorCode          <- 0
lReturn             <- list(
  TreatmentID = vSelectedTreatments, 
  AllocRatio  = vAllocationRatio,
  ErrorCode   = nErrorCode
)
return(lReturn)

# Example 2: Assuming the allocation in the second part of the trial is 1:1:2 for Control:Experimental 1:Experimental 2
vSelectedTreatments <- c(1, 2)  # Experimental 2 will receive twice as many participants as Experimental 1 or Control. 
vAllocationRatio    <- c(1, 2)
nErrorCode          <- 0
lReturn             <- list(
  TreatmentID = vSelectedTreatments, 
  AllocRatio  = vAllocationRatio,
  ErrorCode   = nErrorCode
)
return(lReturn)
}
