% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePatientOutcomePercentAtZero.R
\name{SimulatePatientOutcomePercentAtZero}
\alias{SimulatePatientOutcomePercentAtZero}
\title{Simulate patient outcomes from a normal distribution with a percent of patients having an outcome of 0.}
\usage{
SimulatePatientOutcomePercentAtZero(
  NumSub,
  TreatmentID,
  Mean,
  StdDev,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{The number of subjects that need to be simulated, integer value}

\item{TreatmentID}{A vector of treatment ids, 0 = treatment 1, 1 = Treatment 2. length( TreatmentID ) = NumSub}

\item{Mean}{A vector of length = 2 with the means of the two treatments.}

\item{StdDev}{A vector of length = 2 with the standard deviations of each treatment}

\item{UserParam}{A list of user defined parameters in East.   The default must be NULL resulting in ignoring the percent of patients at 0.
If UseParam is supplied, the list must contain the following named elements:
\describe{
   \item{UserParam$dProbOfZeroOutcomeCtrl}{A value in (0, 1) that defines the probability a patient will have an outcome of 0 on the control (ctrl) treatment.}
   \item{UserParam$dProbOfZeroOutcomeExp}{A value in (0, 1) that defines the probability a patient will have an outcome of 0 on the control (ctrl) treatment.}
   }}
}
\description{
In this example, the continuous outcome is a patient's change from baseline.   For this function, 20% of patients are believed to have no change due to treatment.  
As such, this function simulations patient outcome where, on average, 20% will have a value of 0 for the outcome and 80%, on average, will have their value
simulated from a normal distribution with the mean and standard deviation as sent from East.
}
