% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePatientSurvivalMixtureExponentials.R
\name{SimulatePatientSurvivalMixtureExponentials}
\alias{SimulatePatientSurvivalMixtureExponentials}
\title{This function simulates from exponential, just included as a simple example as a starting point}
\usage{
SimulatePatientSurvivalMixtureExponentials(
  NumSub,
  NumArm,
  TreatmentID,
  SurvMethod,
  NumPrd,
  PrdTime,
  SurvParam,
  UserParam = NULL
)
}
\description{
This function simulates from exponential, just included as a simple example as a starting point
}
