% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateTTEPatientSurvivalMixtureExponentials.R
\name{SimulatePatientSurvivalMixtureExponentials}
\alias{SimulatePatientSurvivalMixtureExponentials}
\title{Simulate patient outcomes from a Mixture Exponential distribution}
\usage{
SimulatePatientSurvivalMixtureExponentials(
  NumSub,
  NumArm,
  TreatmentID,
  SurvMethod,
  NumPrd,
  PrdTime,
  SurvParam,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{The number of patient times to generate for the trial. This is a single numeric value, e.g., 250.}

\item{NumArm}{The number of arms in the trial, a single numeric value. For a two-arm trial, this will be 2.}

\item{TreatmentID}{A vector of treatment IDs, 0 = treatment 1, 1 = treatment 2. The length of TreatmentID must equal NumSub.}

\item{SurvMethod}{This value is pulled from the Input Method drop-down list. It will be 1 (Hazard Rate), 2 (Cumulative % survival), or 3 (Medians).}

\item{NumPrd}{The number of time periods that are provided.}

\item{UserParam}{A list of user-defined parameters in East or East Horizon. The default is NULL. If UserParam is supplied, 
it must contain the following:
\describe{
     \item{UserParam$QtyOfSubgroups}{The quantity of patient subgroups. For each subgroup II = 1,2,...,QtyOfSubgroups, 
     you must specify ProbSubgroupII, MedianTTECtrlSubgroupII, and MedianTTEExpSubgroupII.}
     \item{UserParam$ProbSubgroup1}{The probability a patient is in subgroup 1.}
     \item{UserParam$MedianTTECtrlSubgroup1}{The median time-to-event for a patient in subgroup 1 that receives control treatment.}
     \item{UserParam$MedianTTEExpSubgroup1}{The median time-to-event for a patient in subgroup 1 that receives experimental treatment.}
     \item{UserParam$ProbSubgroup2}{The probability a patient is in subgroup 2.}
     \item{UserParam$MedianTTECtrlSubgroup2}{The median time-to-event for a patient in subgroup 2 that receives control treatment.}
     \item{UserParam$MedianTTEExpSubgroup2}{The median time-to-event for a patient in subgroup 2 that receives experimental treatment.}
}}
}
\description{
This function simulates patient data from a mixture of Exponential distributions. The mixture is based on patient subgroups.  
For each subgroup, you specify the median time-to-event for the control and experimental treatments as well as the probability 
a patient belongs in a specific group. The required function signature for integration with East or East Horizon includes the SurvMethod, 
NumPrd, PrdTime, and SurvParam, which are ignored in this function, and only the parameters in UserParam are utilized.
}
