% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePatientSurvivalWeibull.R
\name{SimulatePatientSurvivalWeibull}
\alias{SimulatePatientSurvivalWeibull}
\title{Simulate time-to-event patient data from a Weibull distribution}
\usage{
SimulatePatientSurvivalWeibull(
  NumSub,
  NumArm,
  TreatmentID,
  SurvMethod,
  NumPrd,
  PrdTime,
  SurvParam,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{NumSub - The number of patients (subjects) in the trial.  NumSub survival times need to be generated for the trial.
This is a single numeric value, eg 250.}

\item{NumArm}{The number of arms in the trial, a single numeric value.  For a two arm trial, this will be 2. 
The SurvParam depends on input in East. In the simulation window on the Response Generation tab}
}
\description{
Simulate time-to-event patient data from a Weibull distribution
}
