% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePatientSurvivalWeibull.R
\name{SimulatePatientSurvivalWeibull}
\alias{SimulatePatientSurvivalWeibull}
\title{Simulate patient outcomes from a Weibull distribution}
\usage{
SimulatePatientSurvivalWeibull(
  NumSub,
  NumArm,
  TreatmentID,
  SurvMethod,
  NumPrd,
  PrdTime,
  SurvParam,
  UserParam = NULL
)
}
\arguments{
\item{NumSub}{The number of patient times to generate for the trial. This is a single numeric value, e.g., 250.}

\item{NumArm}{The number of arms in the trial, a single numeric value. For a two-arm trial, this will be 2.}

\item{TreatmentID}{A vector of treatment IDs. `0 = treatment 1`, `1 = treatment 2`. The length of `TreatmentID` must be equal to `NumSub`.}

\item{NumPrd}{Number of time periods that are provided.}

\item{UserParam}{A list of user-defined parameters in East or East Horizon. The default is `NULL`.
If `UserParam` is supplied, it must contain the following:
\describe{
   \item{`UserParam$dShapeCtrl`}{The shape parameter in the Weibull distribution for the control treatment.}  
   \item{`UserParam$dScaleCtrl`}{The scale parameter in the Weibull distribution for the control treatment.}
   \item{`UserParam$dShapeExp`}{The shape parameter in the Weibull distribution for the experimental treatment.}  
   \item{`UserParam$dScaleExp`}{The scale parameter in the Weibull distribution for the experimental treatment.}
}}
}
\value{
A list with the following components:
\item{`SurvivalTime`}{A vector of simulated survival times for patients.}
\item{`ErrorCode`}{An integer error code (`0` if no error, `-1` if user parameters are missing).}
}
\description{
This function simulates patient data from a Weibull (shape, scale) distribution. The `rweibull` function in the `stats` package
is used to simulate the survival time. See help on `rweibull`.  

The required function signature for integration with East or East Horizon includes the `SurvMethod`, `NumPrd`, `PrdTime`, and `SurvParam`, 
which are ignored in this function, and only the parameters in `UserParam` are utilized.
}
